split Internal.hs module
